+++ packages/service-core/src/entry/entry-index.ts
@@ -3,3 +3,4 @@ export * from './commands/config-command.js';
 export * from './commands/migrate-action.js';
 export * from './commands/start-action.js';
 export * from './commands/teardown-action.js';
+export * from './commands/compact-action.js';